Removes code related to s2i based URLs
