+++ pkg/url/url_test.go
@@ -1,176 +1,18 @@
 package url
 
 import (
-	"fmt"
-	"reflect"
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	kappsv1 "k8s.io/api/apps/v1"
-
-	routev1 "github.com/openshift/api/route/v1"
-	applabels "github.com/openshift/odo/pkg/application/labels"
-	componentlabels "github.com/openshift/odo/pkg/component/labels"
 	"github.com/openshift/odo/pkg/kclient"
 	"github.com/openshift/odo/pkg/kclient/fake"
 	"github.com/openshift/odo/pkg/localConfigProvider"
 	"github.com/openshift/odo/pkg/occlient"
 	"github.com/openshift/odo/pkg/testingutil"
-	"github.com/openshift/odo/pkg/url/labels"
-	"github.com/openshift/odo/pkg/version"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	kappsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	ktesting "k8s.io/client-go/testing"
 )
-
-func TestExists(t *testing.T) {
-	tests := []struct {
-		name            string
-		urlName         string
-		componentName   string
-		applicationName string
-		wantBool        bool
-		routes          routev1.RouteList
-		labelSelector   string
-		wantErr         bool
-	}{
-		{
-			name:            "correct values and Host found",
-			urlName:         "nodejs",
-			componentName:   "nodejs",
-			applicationName: "app",
-			routes: routev1.RouteList{
-				Items: []routev1.Route{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "nodejs",
-							Labels: map[string]string{
-								applabels.ApplicationLabel:     "app",
-								componentlabels.ComponentLabel: "nodejs",
-								applabels.ManagedBy:            "odo",
-								applabels.ManagerVersion:       version.VERSION,
-								labels.URLLabel:                "nodejs",
-							},
-						},
-						Spec: routev1.RouteSpec{
-							To: routev1.RouteTargetReference{
-								Kind: "Service",
-								Name: "nodejs-app",
-							},
-							Port: &routev1.RoutePort{
-								TargetPort: intstr.FromInt(8080),
-							},
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "wildfly",
-							Labels: map[string]string{
-								applabels.ApplicationLabel:     "app",
-								componentlabels.ComponentLabel: "wildfly",
-								applabels.ManagedBy:            "odo",
-								applabels.ManagerVersion:       version.VERSION,
-								labels.URLLabel:                "wildfly",
-							},
-						},
-						Spec: routev1.RouteSpec{
-							To: routev1.RouteTargetReference{
-								Kind: "Service",
-								Name: "wildfly-app",
-							},
-							Port: &routev1.RoutePort{
-								TargetPort: intstr.FromInt(9100),
-							},
-						},
-					},
-				},
-			},
-			wantBool:      true,
-			labelSelector: "app.kubernetes.io/instance=nodejs,app.kubernetes.io/part-of=app",
-			wantErr:       false,
-		},
-		{
-			name:            "correct values and Host not found",
-			urlName:         "example",
-			componentName:   "nodejs",
-			applicationName: "app",
-			routes: routev1.RouteList{
-				Items: []routev1.Route{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "nodejs",
-							Labels: map[string]string{
-								applabels.ApplicationLabel:     "app",
-								componentlabels.ComponentLabel: "nodejs",
-								applabels.ManagedBy:            "odo",
-								applabels.ManagerVersion:       version.VERSION,
-								labels.URLLabel:                "nodejs",
-							},
-						},
-						Spec: routev1.RouteSpec{
-							To: routev1.RouteTargetReference{
-								Kind: "Service",
-								Name: "nodejs-app",
-							},
-							Port: &routev1.RoutePort{
-								TargetPort: intstr.FromInt(8080),
-							},
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "wildfly",
-							Labels: map[string]string{
-								applabels.ApplicationLabel:     "app",
-								componentlabels.ComponentLabel: "wildfly",
-								applabels.ManagedBy:            "odo",
-								applabels.ManagerVersion:       version.VERSION,
-								labels.URLLabel:                "wildfly",
-							},
-						},
-						Spec: routev1.RouteSpec{
-							To: routev1.RouteTargetReference{
-								Kind: "Service",
-								Name: "wildfly-app",
-							},
-							Port: &routev1.RoutePort{
-								TargetPort: intstr.FromInt(9100),
-							},
-						},
-					},
-				},
-			},
-			wantBool:      false,
-			labelSelector: "app.kubernetes.io/instance=nodejs,app.kubernetes.io/part-of=app",
-			wantErr:       false,
-		},
-	}
-	for _, tt := range tests {
-		client, fakeClientSet := occlient.FakeNew()
-
-		fakeClientSet.RouteClientset.PrependReactor("list", "routes", func(action ktesting.Action) (bool, runtime.Object, error) {
-			if !reflect.DeepEqual(action.(ktesting.ListAction).GetListRestrictions().Labels.String(), tt.labelSelector) {
-				return true, nil, fmt.Errorf("labels not matching with expected values, expected:%s, got:%s", tt.labelSelector, action.(ktesting.ListAction).GetListRestrictions())
-			}
-			return true, &tt.routes, nil
-		})
-
-		exists, err := Exists(client, tt.urlName, tt.componentName, tt.applicationName)
-		if err == nil && !tt.wantErr {
-			if (len(fakeClientSet.RouteClientset.Actions()) != 1) && (tt.wantErr != true) {
-				t.Errorf("expected 1 action in ListRoutes got: %v", fakeClientSet.RouteClientset.Actions())
-			}
-			if exists != tt.wantBool {
-				t.Errorf("expected exists to be:%t, got :%t", tt.wantBool, exists)
-			}
-		} else if err == nil && tt.wantErr {
-			t.Errorf("test failed, expected: %s, got %s", "false", "true")
-		} else if err != nil && !tt.wantErr {
-			t.Errorf("test failed, expected: %s, got %s", "no error", "error:"+err.Error())
-		}
-	}
-}
 
 func TestPush(t *testing.T) {
 	type deleteParameters struct {
@@ -834,154 +676,6 @@ func TestPush(t *testing.T) {
 				IsRouteSupported: tt.args.isRouteSupported,
 			}); (err != nil) != tt.wantErr {
 				t.Errorf("Push() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestConvertEnvinfoURL(t *testing.T) {
-	serviceName := "testService"
-	urlName := "testURL"
-	host := "com"
-	secretName := "test-tls-secret"
-	tests := []struct {
-		name       string
-		envInfoURL localConfigProvider.LocalURL
-		wantURL    URL
-	}{
-		{
-			name: "Case 1: insecure URL",
-			envInfoURL: localConfigProvider.LocalURL{
-				Name:   urlName,
-				Host:   host,
-				Port:   8080,
-				Secure: false,
-				Kind:   localConfigProvider.INGRESS,
-			},
-			wantURL: URL{
-				TypeMeta:   metav1.TypeMeta{Kind: "URL", APIVersion: "odo.dev/v1alpha1"},
-				ObjectMeta: metav1.ObjectMeta{Name: urlName},
-				Spec:       URLSpec{Host: fmt.Sprintf("%s.%s", urlName, host), Port: 8080, Secure: false, Kind: localConfigProvider.INGRESS},
-			},
-		},
-		{
-			name: "Case 2: secure Ingress URL without tls secret defined",
-			envInfoURL: localConfigProvider.LocalURL{
-				Name:   urlName,
-				Host:   host,
-				Port:   8080,
-				Secure: true,
-				Kind:   localConfigProvider.INGRESS,
-			},
-			wantURL: URL{
-				TypeMeta:   metav1.TypeMeta{Kind: "URL", APIVersion: "odo.dev/v1alpha1"},
-				ObjectMeta: metav1.ObjectMeta{Name: urlName},
-				Spec:       URLSpec{Host: fmt.Sprintf("%s.%s", urlName, host), Port: 8080, Secure: true, TLSSecret: fmt.Sprintf("%s-tlssecret", serviceName), Kind: localConfigProvider.INGRESS},
-			},
-		},
-		{
-			name: "Case 3: secure Ingress URL with tls secret defined",
-			envInfoURL: localConfigProvider.LocalURL{
-				Name:      urlName,
-				Host:      host,
-				Port:      8080,
-				Secure:    true,
-				TLSSecret: secretName,
-				Kind:      localConfigProvider.INGRESS,
-			},
-			wantURL: URL{
-				TypeMeta:   metav1.TypeMeta{Kind: "URL", APIVersion: "odo.dev/v1alpha1"},
-				ObjectMeta: metav1.ObjectMeta{Name: urlName},
-				Spec:       URLSpec{Host: fmt.Sprintf("%s.%s", urlName, host), Port: 8080, Secure: true, TLSSecret: secretName, Kind: localConfigProvider.INGRESS},
-			},
-		},
-		{
-			name: "Case 4: Insecure route URL",
-			envInfoURL: localConfigProvider.LocalURL{
-				Name: urlName,
-				Port: 8080,
-				Kind: localConfigProvider.ROUTE,
-			},
-			wantURL: URL{
-				TypeMeta:   metav1.TypeMeta{Kind: "URL", APIVersion: "odo.dev/v1alpha1"},
-				ObjectMeta: metav1.ObjectMeta{Name: urlName},
-				Spec:       URLSpec{Port: 8080, Secure: false, Kind: localConfigProvider.ROUTE},
-			},
-		},
-		{
-			name: "Case 4: Secure route URL",
-			envInfoURL: localConfigProvider.LocalURL{
-				Name:   urlName,
-				Port:   8080,
-				Secure: true,
-				Kind:   localConfigProvider.ROUTE,
-			},
-			wantURL: URL{
-				TypeMeta:   metav1.TypeMeta{Kind: "URL", APIVersion: "odo.dev/v1alpha1"},
-				ObjectMeta: metav1.ObjectMeta{Name: urlName},
-				Spec:       URLSpec{Port: 8080, Secure: true, Kind: localConfigProvider.ROUTE},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			url := NewURLFromEnvinfoURL(tt.envInfoURL, serviceName)
-			if !reflect.DeepEqual(url, tt.wantURL) {
-				t.Errorf("Expected %v, got %v", tt.wantURL, url)
-			}
-		})
-	}
-}
-
-func TestGetURLString(t *testing.T) {
-	cases := []struct {
-		name          string
-		protocol      string
-		URL           string
-		ingressDomain string
-		isS2I         bool
-		expected      string
-	}{
-		{
-			name:          "simple s2i case",
-			protocol:      "http",
-			URL:           "example.com",
-			ingressDomain: "",
-			isS2I:         true,
-			expected:      "http://example.com",
-		},
-		{
-			name:          "all blank with s2i",
-			protocol:      "",
-			URL:           "",
-			ingressDomain: "",
-			isS2I:         true,
-			expected:      "",
-		},
-		{
-			name:          "all blank without s2i",
-			protocol:      "",
-			URL:           "",
-			ingressDomain: "",
-			isS2I:         false,
-			expected:      "",
-		},
-		{
-			name:          "devfile case",
-			protocol:      "http",
-			URL:           "",
-			ingressDomain: "spring-8080.192.168.39.247.nip.io",
-			isS2I:         false,
-			expected:      "http://spring-8080.192.168.39.247.nip.io",
-		},
-	}
-
-	for _, testCase := range cases {
-		t.Run(testCase.name, func(t *testing.T) {
-			output := GetURLString(testCase.protocol, testCase.URL, testCase.ingressDomain, testCase.isS2I)
-			if output != testCase.expected {
-				t.Errorf("Expected: %v, got %v", testCase.expected, output)
-
 			}
 		})
 	}
